\name{SpatDataFrame-class}

\docType{class}

\alias{SpatDataFrame}
\alias{SpatDataFrame-class}
\alias{Rcpp_SpatDataFrame-class}
\alias{SpatOptions}
\alias{SpatOptions-class}
\alias{Rcpp_SpatOptions-class}
\alias{SpatMessages}
\alias{SpatMessages-class}
\alias{Rcpp_SpatMessages-class}
\alias{SpatRasterCollection}
\alias{SpatRasterCollection-class}
\alias{Rcpp_SpatRasterCollection-class}
\alias{SpatStack}
\alias{SpatStack-class}
\alias{Rcpp_SpatStack-class}
\alias{show,SpatStack-method}


\alias{show_messages}
\alias{show,SpatDataFrame-method}

\title{C++ classes}

\description{
C++ classes and methods. Not for end-users.
}



\keyword{classes}
\keyword{spatial}
