\name{deprecated}

\alias{gridDistance}
\alias{gridDistance,SpatRaster-method}


\title{deprecated methods}

\description{
These methods are deprecated and will be removed in future versions
}

\usage{
\S4method{gridDistance}{SpatRaster}(x, ...)
}

\arguments{
  \item{x}{object}
  \item{...}{additional arguments}
}


