\name{geomtype}

\docType{methods}

\alias{geomtype}
\alias{geomtype,SpatLayer-method}

\title{Geometry type of a SpatLayer}

\description{
Get the geometry type (points, lines, or polygons) of a SpatLayer
} 

\usage{
\S4method{geomtype}{SpatLayer}(x, ...)
}

\arguments{
\item{x}{SpatLayer}
\item{...}{additional arguments. None implemented }  
}

\value{character}

\keyword{methods}
\keyword{spatial}

