\name{meta}

\docType{methods}

\alias{meta}
\alias{meta,SpatRaster-method}

\title{meta}

\description{ 
Get metadata associated with the sources or layers of a SpatRaster 
}

\usage{
\S4method{meta}{SpatRaster}(x, layers=FALSE)
}

\arguments{
  \item{x}{SpatRaster}
  \item{layers}{logical. Should the layer level metadata be returned?}
}

\value{
list
}


\keyword{methods}
\keyword{spatial}
