\name{metadata}

\docType{methods}

\alias{metadata}
\alias{metadata,SpatRaster-method}

\title{metadata}

\description{ 
Return metadata associated with the layers of a SpatRaster 
}

\usage{
\S4method{metadata}{SpatRaster}(x)
}

\arguments{
  \item{x}{SpatRaster or SpatRasterDataset}
}

\value{
list
}


\keyword{methods}
\keyword{spatial}
