\name{SpatDataFrame-class}

\docType{class}

\alias{SpatDataFrame}
\alias{SpatDataFrame-class}
\alias{Rcpp_SpatDataFrame-class}

\alias{show,SpatDataFrame-method}

\title{Class "SpatDataFrame" }

\description{
Objects of class SpatDataFrame. 
}


\section{Slots}{
	 \describe{
    \item{\code{ptr}:}{pointer to the C++ class} 
  }
}

\section{Methods}{
  \describe{
  \item{show}{display the structure of a SpatDataFrame}
 }
}



\keyword{classes}
\keyword{spatial}
