\name{SpatLayer-class}

\docType{class}

\alias{SpatLayer}
\alias{SpatLayer-class}
\alias{Rcpp_SpatLayer-class}

\alias{show,SpatLayer-method}

\title{Class "SpatLayer" }

\description{
Objects of class SpatLayer. 
}

\section{Objects from the Class}{
You can use the \code{\link{vect}} method to create SpatLayer objects.
}

\section{Slots}{
	 \describe{
    \item{\code{ptr}:}{pointer to the C++ class} 
  }
}

\section{Methods}{
  \describe{
  \item{show}{display values of a SpatLayer}
 }
}



\keyword{classes}
\keyword{spatial}
