\name{SpatMessages-class}

\docType{class}

\alias{SpatMessages}
\alias{SpatMessages-class}
\alias{Rcpp_SpatMessages-class}
\alias{show_messages}


\title{Messages}

\description{
Class and method for showing warning and error messages. For developers, not for end-users.
}

\section{Slots}{
	 \describe{
    \item{\code{ptr}:}{pointer to the C++ class} 
  }
}


\keyword{classes}
\keyword{spatial}
