\name{subClass}

\docType{methods}

\alias{subClass}
\alias{subClass,SpatVector-method}

\title{Get the sub-class of a SpatVector object}

\description{
et the sub-class (points, lines, or polygons) of a SpatialVector object
} 

\usage{
\S4method{subClass}{SpatVector}(x, ...)
}

\arguments{
\item{x}{SpatialVector}
\item{...}{additional arguments. None implemented }  
}


\value{character}



\keyword{methods}
\keyword{spatial}

