\name{vrt_tiles}


\alias{vrt_tiles} 

\title{filenames of VRT tiles}

\description{ 
Get the filenames of the tiles in a Virtual Raster Dataset (VRT) 
}

\usage{
vrt_tiles(x)
}

\arguments{
  \item{x}{character (filename) or SpatRaster}
}

\value{
character
}

\seealso{
\code{\link{vrt}}
}


\keyword{methods}
\keyword{spatial}
