\name{colors}

\docType{methods}

\alias{coltab}
\alias{coltab,SpatRaster-method}
  
\title{Color table}

\description{
Get the color table(s) associated with a SpatRaster if there are any
}

\usage{
\S4method{coltab}{SpatRaster}(x)
}


\arguments{
  \item{x}{SpatRaster}
}  


\value{
data.frame
}


\keyword{methods}
\keyword{spatial}

