\name{deprecated}

\alias{gridDistance}
\alias{gridDistance,SpatRaster-method}


\title{deprecated methods}

\description{
This method is no longer available. Use \code{\link{gridDist}} instead
}

\usage{
\S4method{gridDistance}{SpatRaster}(x, ...)
}

\arguments{
  \item{x}{object}
  \item{...}{additional arguments}
}


