\name{deprecated}

\alias{setCats}
\alias{setCats,SpatRaster-method}
\alias{area}
\alias{area,SpatRaster-method}

\title{deprecated methods}

\description{
These methods are deprecated and will be removed in future versions
}

\usage{
\S4method{setCats}{SpatRaster}(x, ...)

\S4method{area}{SpatRaster}(x, ...)
}

\arguments{
  \item{x}{object}
  \item{...}{additional arguments}
}


