\name{geomtype}

\docType{methods}

\alias{geomtype}
\alias{geomtype,SpatVector-method}
\alias{geomtype,Spatial-method}
\alias{is.points}
\alias{is.lines}
\alias{is.polygons}
\alias{is.points,SpatVector-method}
\alias{is.lines,SpatVector-method}
\alias{is.polygons,SpatVector-method}

\title{Geometry type of a SpatVector}

\description{
Get the geometry type (points, lines, or polygons) of a SpatVector
} 

\usage{
\S4method{geomtype}{SpatVector}(x, ...)

\S4method{is.points}{SpatVector}(x, ...)

\S4method{is.lines}{SpatVector}(x, ...)

\S4method{is.polygons}{SpatVector}(x, ...)
}

\arguments{
\item{x}{SpatVector}
\item{...}{additional arguments. None implemented }  
}

\value{character}

\keyword{methods}
\keyword{spatial}

