\name{geomtype}

\docType{methods}

\alias{geomtype}
\alias{geomtype,SpatVector-method}

\title{Geometry type of a SpatVector}

\description{
Get the geometry type (points, lines, or polygons) of a SpatVector
} 

\usage{
\S4method{geomtype}{SpatVector}(x, ...)
}

\arguments{
\item{x}{SpatVector}
\item{...}{additional arguments. None implemented }  
}

\value{character}

\keyword{methods}
\keyword{spatial}

