\name{plot}

\alias{plot}


\title{Make a map with "terra"}

\description{

Choose one of the following help files for plotting a \code{\link[=plot,SpatRaster,missing-method]{SpatRaster}}, 
\code{\link[=plot,SpatVector,missing-method]{SpatVector}}, \code{\link[=plot,SpatExtent,missing-method]{SpatExtent}} or \code{\link[=plot,SpatGraticule,missing-method]{SpatGraticule}}.

}

\seealso{
 \code{\link{points}}, \code{\link{lines}}, \code{\link{polys}}, \code{\link{image}}, \code{scatter\link[terra:scatter]{plot}}, 
 scale bar: \code{\link{sbar}}, north arrow: \code{\link{north}}
}

\keyword{methods}
\keyword{spatial}

