\name{Rcpp-class}

\docType{class}

\alias{SpatGeomRing}
\alias{SpatGeomRings}
\alias{SpatPolygons}

\alias{SpatGeomRing-class}
\alias{SpatGeomRings-class}
\alias{SpatPolygons-class}

\alias{Rcpp_SpatGeomRing-class}
\alias{Rcpp_SpatGeomRings-class}
\alias{Rcpp_SpatPolygons-class}

\title{ Rcpp classes}

\description{
These classes are for internal use only
}


\keyword{classes}
\keyword{spatial}
