\name{SpatVector-class}

\docType{class}

\alias{SpatVector}
\alias{SpatVector-class}
\alias{Rcpp_SpatVector-class}

\alias{show,SpatVector-method}

\title{Class "SpatVector" }

\description{
Objects of class SpatVector. 
}

\section{Objects from the Class}{
You can use the \code{\link{vect}} method to create SpatVector objects.
}

\section{Slots}{
	 \describe{
    \item{\code{ptr}:}{pointer to the C++ class} 
  }
}

\section{Methods}{
  \describe{
  \item{show}{display values of a SpatVector}
 }
}



\keyword{classes}
\keyword{spatial}
