\name{SpatVector-class}

\docType{class}

\alias{SpatVector}
\alias{SpatVector-class}

\alias{SpatVectorCollection}
\alias{SpatVectorCollection-class}

\alias{SpatVectorProxy}
\alias{SpatVectorProxy-class}

\alias{Rcpp_SpatVector-class}
\alias{PackedSpatVector-class}

\alias{show,SpatVector-method}

\title{Class "SpatVector" }

\description{
\code{SpatVector} can represent points, lines or polygons.

\code{SpatVectorCollection} can hold a collection of SpatVectors

\code{SpatVectorProxy} is a SpatVector for which the data are on-disk in-stead of in memory. 
}

\keyword{classes}
\keyword{spatial}
