\name{writeRaster}

\alias{writeRaster,SpatRaster,character-method}

\alias{writeRaster}

\title{Write raster data to a file}

\description{
Write a SpatRaster object to a file. 
}

\usage{
\S4method{writeRaster}{SpatRaster,character}(x, filename, overwrite=FALSE, ...)
}

\arguments{
\item{x}{Raster* object}
\item{filename}{Output filename}
\item{overwrite}{logical}
\item{...}{Additional arguments (to be done) }
}

\value{
This function is used for the side-effect of writing values to a file.
}


\keyword{ spatial }
\keyword{ methods }

